Share agent execution selection helpers
