+++ common/__tests__/execution-defaults.test.js
@@ -1,0 +1,43 @@
+import { describe, expect, it } from 'bun:test';
+import {
+  executionDefaultsForAgent,
+  normalizeSupportedPermissionMode,
+  normalizeSupportedThinkingMode,
+} from '../execution-defaults.js';
+
+describe('execution defaults', () => {
+  it('merges one agent override onto global defaults', () => {
+    const globalSettings = { ownerId: 'codex', schemaVersion: 1, values: { source: 'global' } };
+    const agentSettings = { ownerId: 'codex', schemaVersion: 1, values: { source: 'agent' } };
+    expect(executionDefaultsForAgent({
+      global: {
+        permissionMode: 'default',
+        thinkingMode: 'low',
+        agentSettingsById: { codex: globalSettings },
+      },
+      byAgent: {
+        codex: {
+          permissionMode: 'acceptEdits',
+          agentSettingsById: { codex: agentSettings },
+        },
+      },
+    }, 'codex')).toEqual({
+      permissionMode: 'acceptEdits',
+      thinkingMode: 'low',
+      agentSettingsById: { codex: agentSettings },
+    });
+  });
+
+  it('falls back to canonical modes without configured defaults', () => {
+    expect(executionDefaultsForAgent(null, 'codex')).toEqual({
+      permissionMode: 'default',
+      thinkingMode: 'none',
+      agentSettingsById: {},
+    });
+  });
+
+  it('selects the first supported mode when a configured mode is unavailable', () => {
+    expect(normalizeSupportedPermissionMode('plan', ['default', 'acceptEdits'])).toBe('default');
+    expect(normalizeSupportedThinkingMode('ultra', ['low', 'high'])).toBe('low');
+  });
+});
