+++ common/package.json
@@ -7,6 +7,7 @@
     "./agent-auth": "./agent-auth.ts",
     "./agent-execution": "./agent-execution.ts",
     "./agent-integration": "./agent-integration.ts",
+    "./agent-settings": "./agent-settings.ts",
     "./agents": "./agents.ts",
     "./api-provider-templates": "./api-provider-templates.ts",
     "./api-providers": "./api-providers.ts",
@@ -20,6 +21,8 @@
     "./chat-title-contracts": "./chat-title-contracts.ts",
     "./chat-types": "./chat-types.ts",
     "./chat-view": "./chat-view.ts",
+    "./client-chat-id": "./client-chat-id.ts",
+    "./execution-defaults": "./execution-defaults.ts",
     "./file-contracts": "./file-contracts.ts",
     "./generation-test-contracts": "./generation-test-contracts.ts",
     "./gh": "./gh.ts",
