+++ web/src/lib/chat/conversation/conversation-slash-command-service.ts
@@ -17,7 +17,7 @@ import {
 	parseSteerCommand,
 	parseTagCommand,
 } from '$lib/chat/composer/slash-commands.js';
-import { createClientChatId } from '$lib/chat/sessions/client-chat-id.js';
+import { createClientChatId } from '$shared/client-chat-id';
 import { createClientCommandId } from '$lib/chat/conversation/client-command-id.js';
 import type {
 	ScheduleInCommandError,
