+++ web/src/lib/components/chat/NewChatDialog.svelte
@@ -5,7 +5,7 @@
 
 	import * as Dialog from '$lib/components/ui/dialog';
 	import { getAppShell, getChatSessions, getWorkspaceCoordinator } from '$lib/context';
-	import { createClientChatId } from '$lib/chat/sessions/client-chat-id.js';
+	import { createClientChatId } from '$shared/client-chat-id';
 	import { gotoChat } from '$lib/chat/actions/chat-navigation.js';
 	import type { NewChatConfig } from '$lib/types/app';
 	import NewChatForm from './NewChatForm.svelte';
