+++ web/src/lib/components/sidebar/sidebar-controller.svelte.ts
@@ -12,7 +12,7 @@ import {
 	setChatTags,
 	updateChatProjectPath,
 } from '$lib/api/chats.js';
-import { createClientChatId } from '$lib/chat/sessions/client-chat-id.js';
+import { createClientChatId } from '$shared/client-chat-id';
 import type { ProjectPathPatchResponse } from '$shared/chat-command-contracts';
 import type { ChatSessionRecord } from '$lib/types/chat-session';
 import type { ChatListEntry } from '$shared/chat-list';
